moved bookmarks list item to common-ui
