+++ app/src/main/res/layout/row_bookmark_two_line_item.xml
@@ -15,7 +15,7 @@
   -->
 
 
-<com.duckduckgo.app.bookmarks.ui.BookmarkTwoLineListItem
+<com.duckduckgo.common.ui.view.listitem.BookmarksListItem
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
